Sync and Fetch separated into custom hook files
